requirements.txt
aiogram==3.10.0
telethon==1.34.0
aiohttp==3.9.5
python-dotenv==1.0.1